SaveManager + GlobalDataStore
The SaveManager lives in the GlobalDataStore and handles all player save data which is stored in the class SaveData. This includes saving and loading. The GlobalDataStore stores all scence shared states and objects like the SaveManager. The player save data class SaveData and the GlobalDataStore needs to be extended as it is used.