SettingsManager + Extension SaveManger
The SettingsManger stores all Settings which are global,not linked to a single save like audio settings, in the structure SettingsData which lives in the GlobalDataStore. Only 1 SettingsFile (settings.dat) can exist. It is now possible to peek(view) saves without loading them as the active save. A save loaded in a previous session will always be loaded in the next instead of the default save.