Player prefab rework + MenuManager fix
The player prefab was reworked to allow for easier replacing of the PlayerModell and an overlay for UI Elemts like the healthbar was added with an canvas. The Contents of the Player overlay have not been made. The MenuManager only partially covered part of the canvas, instead of covering the full 1920*1080 px size of the canvas. This has now been fixed. It now covers the whole 1920*1080 px size canvas.